pyproject.toml
[project]
name = "pypi-fairylandfuture"
version = "1.3.3"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Python
    "pip-tools", # pip-sync, pip-compile
    "pipreqs", # pipreqs --encoding=UTF-8 --force .
    "pipdeptree", # pipdeptree -fl
    "pip-review", # pip-review --interactive
    "pip-autoremove", # pip-autoremove -y <package name>
    "pigar", # pigar -p requirements.txt
    "setuptools",
    "wheel",
    "pip",
    "uv",
    # Code Style
    "black", # black --line-length 170 --target-version py311 .
    "flake8",
    "isort",
    "mypy",
    "pydocstyle",
    "pylint",
    "autopep8",
    "pylama",
    # Tools
    "loguru", # Log handles
    "python-dateutil", # Date handles
    "python-dotenv", # Environment variable handles
    "requests", # HTTP(S) requests
    "pyyaml", # YAML handles
    "cryptography", # required C library, SSL/TLS, encryption handles
    "netifaces", # required C library, network interface handles  ### PEP517
    "pymysql", # MySQL
    "psycopg2", # PostgreSQL
    "psycopg2-binary", # PostgreSQL
    "mysql-connector-python", # MySQL
    "sqlalchemy", # SQLAlchemy ORM
    "DBUtils", # Database connection pooling
    "aiomysql", # MySQL async
    "kafka-python", # Kafka
    "faker", # Fake data generation
    "pypinyin", # Chinese Pinyin handles
    "elasticsearch[async]", # Elasticsearch /async
    "opensearch-py", # OpenSearch
    "pydantic", # Data validation and settings management
    "ua-parser",
    "user-agents",
    # Django and related packages
    "django",
    "djangorestframework",
    "django-redis",
    "django-filter",
    "django-cors-headers",
    "django-simpleui",  ### PEP517
    "django-tenant-schemas",  ### PEP517
    "django-tenants",  ### PEP517
    "django-extensions",
    "django-celery-results",
    # API documentation
    "drf-yasg",
    "drf-spectacular",
    "coreapi",
    "pygments",
    "markdown",
    "uritemplate",
    # Auto
    "selenium",
    "webdriver-manager",  ### PEP517
    "beautifulsoup4",
]
[build-system]
requires = [
    "setuptools",
    "wheel",
    "requests",
]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 170
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
    | \.egg-info
    | __pycache__
)/
'''

[tool.isort]
profile = "black"
line_length = 170
skip_gitignore = true
known_first_party = ["fairylandfuture"]

[tool.ruff]
line-length = 170
target-version = "py311"
exclude = [
    ".git",
    ".venv",
    "build",
    "dist",
    "__pycache__",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "E722",  # bare except
]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "build",
    "dist",
    ".venv",
]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"